+++ pkg/msgapi/squish.go
@@ -489,7 +489,7 @@ func (s *Squish) GetMessages() *[]MessageListItem {
 			continue
 		}
 		s.messages = append(s.messages, MessageListItem{
-			MsgNum:      uint32(i + 1),
+			MsgNum:      i + 1,
 			From:        m.From,
 			To:          m.To,
 			Subject:     m.Subject,
